base core
