+++ lib/src/core/base/eden_nav_base_widget.dart
