+++ lib/src/core/base/mixin/eden_refreshable_mixin.dart
@@ -1,35 +1,29 @@
 import 'package:flutter_eden/eden.dart';
 
-abstract class EdenBaseListWidget<T extends EdenBaseListController>
-    extends EdenBaseWidget<T> {
-  EdenBaseListWidget({super.key});
-
-  ///body view
-  Widget buildListView(BuildContext context, T controller);
-
-  ///topView
+mixin EdenRefreshableMixin<T extends EdenBaseListController> {
+  @protected
   Widget renderHeadView(BuildContext context) {
-    return Container();
+    return const SizedBox.shrink();
   }
 
-  @override
-  Widget buildBody(BuildContext context, T controller) {
+  Widget buildListView(BuildContext context, T controller);
+
+  @protected
+  Widget buildRefreshableList(BuildContext context, T controller) {
     return Column(
       children: [
-        renderHeadView(context), //topView
-
+        renderHeadView(context),
         Expanded(
           child: SmartRefresher(
-            enablePullDown: controller.enablePull,
+            enablePullDown: controller.enablePullDown,
             enablePullUp: controller.isNext(),
             controller: controller.refreshController,
             onRefresh: () {
               controller.onRefresh();
             },
-            onLoading: () {
+            onLoading: () async {
               controller.onLoading();
             },
-            primary: false,
             child: buildListView(context, controller),
           ),
         ),
