+++ lib/src/core/eden_core.dart
@@ -1,8 +1,11 @@
 export 'app/eden_material_wrapper.dart';
-export 'eden_base_widget.dart';
-export 'eden_base_list_widget.dart';
-export 'eden_base_service_widget.dart';
-export 'eden_nav_base_widget.dart';
+export 'base/eden_base_service_view.dart';
+export 'base/eden_nav_base_widget.dart';
+
+///base
+export 'base/eden_base_widget.dart';
+export 'base/eden_base_view.dart';
+export 'base/eden_base_list_view.dart';
 
 ///controller
 export 'controllers/eden_base_controller.dart';
